+++ test/unit/validator.test.js
@@ -672,6 +672,37 @@ describe('VALIDATE FUNCTION TESTS ', function () {
     });
 
     differentContentTypesSpecs.forEach((specData) => {
+      it('Should correctly match and validate content type headers having wildcard characters' +
+        ' with collection req/res body', function (done) {
+        let differentContentTypesSpec = fs.readFileSync(specData.path, 'utf-8'),
+          differentContentTypesCollection = fs.readFileSync(path.join(__dirname, VALIDATION_DATA_FOLDER_PATH +
+            '/differentContentTypesCollection.json'), 'utf-8'),
+          resultObj,
+          historyRequest = [],
+          options = {
+            showMissingInSchemaErrors: true,
+            suggestAvailableFixes: true
+          },
+          schemaPack = new Converter.SchemaPack({ type: 'string', data: differentContentTypesSpec }, options);
+
+        getAllTransactions(JSON.parse(differentContentTypesCollection), historyRequest);
+
+        schemaPack.validateTransaction(historyRequest, (err, result) => {
+          expect(err).to.be.null;
+          expect(result).to.be.an('object');
+          resultObj = result.requests[historyRequest[1].id].endpoints[0];
+
+          /**
+           * Both req and res body should have matched content types
+           */
+          expect(resultObj.matched).to.eql(true);
+          expect(resultObj.mismatches).to.have.lengthOf(0);
+          expect(resultObj.responses[_.keys(resultObj.responses)[0]].matched).to.eql(true);
+          expect(resultObj.responses[_.keys(resultObj.responses)[0]].mismatches).to.have.lengthOf(0);
+          done();
+        });
+      });
+
       it('Should correctly match and validate valid json content type with collection req/res body - version:' +
        specData.version, function (done) {
         let differentContentTypesSpec = fs.readFileSync(specData.path, 'utf-8'),
@@ -742,6 +773,31 @@ describe('VALIDATE FUNCTION TESTS ', function () {
     });
 
     multiplePathVarSpecs.forEach((specData) => {
+      it('Should correctly validate path variable in collection that are part of URL itself and are ' +
+        'not present in $request.url.variable', function (done) {
+        let multiplePathVarSpec = fs.readFileSync(specData.path, 'utf-8'),
+          multiplePathVarCollection = fs.readFileSync(path.join(__dirname, VALIDATION_DATA_FOLDER_PATH +
+            '/multiplePathVarCollection.json'), 'utf-8'),
+          resultObj,
+          historyRequest = [],
+          options = {
+            detailedBlobValidation: true,
+            allowUrlPathVarMatching: true
+          },
+          schemaPack = new Converter.SchemaPack({ type: 'string', data: multiplePathVarSpec }, options);
+
+        getAllTransactions(JSON.parse(multiplePathVarCollection), historyRequest);
+
+        schemaPack.validateTransaction(historyRequest, (err, result) => {
+          expect(err).to.be.null;
+          expect(result).to.be.an('object');
+
+          resultObj = result.requests[historyRequest[2].id].endpoints[0];
+          expect(resultObj.mismatches).to.have.lengthOf(0);
+          done();
+        });
+      });
+
       it('Should correctly validate schema having path with various path variables - version: ' +
         specData.version, function (done) {
         let multiplePathVarSpec = fs.readFileSync(specData.path, 'utf-8'),